+++ src/http/index.ts
@@ -1,6 +1,8 @@
 // Barrel: re-exports the http module.
 
 export type * from "./types";
+export * from "./access-point";
+export * from "./insecure-http-warning";
 export * from "./fetch-transport";
 export * from "./mock-transport";
 export * from "./throttle";